backend/requirements.txt
# SpoonOS AI SDK
spoon-ai-sdk>=0.1.0
spoon-toolkits>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# CORS for frontend connection
python-multipart>=0.0.6

# Async support
aiohttp>=3.9.0
